Fix undo and redo commands not working
The problem is that whenever a commit is made to the versioned module planner,
the module planner added into the modulePlannerStateList is not a deep copy. Hence,
even though different commands may be performed, all the module planners stored in
the modulePlannerStateList have the same list of semesters.

The problem was fixed by making sure that the module planner added
to the modulePlannerStateList is a deep copy